refactor(ctrls): move to `pkg`
